gui/requirements.txt
flask
torch
torchvision
ultralytics
pillow
numpy
